tts-service/requirements.txt
chatterbox-tts
torch
torchaudio
fastapi
uvicorn
python-multipart
